# server/internal/docs/swagger.json
{
  "schemes": ["https"],
  "swagger": "2.0",
  "info": {
    "description": "Detailed info about all endpoints",
    "title": "Personal Blog API",
    "contact": {
      "name": "API Support",
      "url": "https://t.me/niyaznassyrov"
    },
    "version": "0.0.1"
  },
  "host": "https://api.nassyrov.net",
  "paths": {
...
    "/user": {
      "get": {
        "description": "This can only be done by the logged-in user.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Get users",
        "parameters": [
          {
            "type": "integer",
            "description": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Offset",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Is Deleted",
            "name": "isDeleted",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Search",
            "name": "search",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.Success"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/response.Error"
            }
          }
        }
      },
      "post": {
        "description": "This can only be done by the logged-in user.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Create user",
        "parameters": [
          {
            "description": "Create User Structure",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userHandler.CreateUserIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.Success"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/response.Error"
            }
          }
        }
      }
    },
    "/user/email/{id}": {
      "put": {
        "description": "This can only be done by the logged-in user.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Update user",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Update User Structure",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userHandler.UpdateUserEmailIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.Success"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/response.Error"
            }
          }
        }
      }
    },
    "/user/password/{id}": {
      "put": {
        "description": "This can only be done by the logged-in user.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Update user",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Update User Structure",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userHandler.UpdateUserPasswordIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.Success"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/response.Error"
            }
          }
        }
      }
    },
    "/user/restore/{id}": {
      "put": {
        "description": "This can only be done by the logged-in user.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Restore user",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.Success"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/response.Error"
            }
          }
        }
      }
    },
    "/user/role/{id}": {
      "put": {
        "description": "This can only be done by the logged-in user.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Update user",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Update User Structure",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userHandler.UpdateUserRoleIn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.Success"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/response.Error"
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "description": "This can only be done by the logged-in user.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Get user",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/userHandler.GetUserOut"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/response.Error"
            }
          }
        }
      },
      "delete": {
        "description": "This can only be done by the logged-in user.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "Delete user",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.Success"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/response.Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "response.Error": {
      "type": "object",
      "properties": {
        "messages": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["error message"]
        },
        "success": {
          "type": "boolean",
          "example": false
        },
        "traceId": {
          "type": "string",
          "example": "string"
        }
      }
    },
    "response.Success": {
      "type": "object",
      "properties": {
        "data": {},
        "success": {
          "type": "boolean",
          "example": true
        },
        "traceId": {
          "type": "string",
          "example": "string"
        }
      }
    },
    "userHandler.CreateUserIn": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
    "userHandler.GetUserOut": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "isDeleted": {
          "type": "boolean"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "userHandler.UpdateUserEmailIn": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "userHandler.UpdateUserPasswordIn": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        }
      }
    },
    "userHandler.UpdateUserRoleIn": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Cookie",
      "in": "header"
    }
  }
}